Extract template tree repository alias to constants
