+++ src/Umbraco.Web.UI.Client/src/packages/templating/templates/tree/constants.ts
@@ -1,1 +1,3 @@
 export * from './tree-item-children/constants.js';
+
+export const UMB_TEMPLATE_TREE_REPOSITORY_ALIAS = 'Umb.Repository.Template.Tree';
