+++ src/Umbraco.Web.UI.Client/src/packages/templating/templates/tree/index.ts
@@ -1,9 +1,5 @@
 export { UmbTemplateTreeRepository } from './template-tree.repository.js';
-export {
-	UMB_TEMPLATE_TREE_REPOSITORY_ALIAS,
-	UMB_TEMPLATE_TREE_STORE_ALIAS,
-	UMB_TEMPLATE_TREE_ALIAS,
-} from './manifests.js';
+export { UMB_TEMPLATE_TREE_STORE_ALIAS, UMB_TEMPLATE_TREE_ALIAS } from './manifests.js';
 export { UMB_TEMPLATE_TREE_STORE_CONTEXT } from './template-tree.store.context-token.js';
 export { type UmbTemplateTreeStore } from './template-tree.store.js';
 export type * from './types.js';
